refactor: 노트에디터 헤더 구조 개선 및 분리
